Use regex for replacement in downloader option
If we use the os shell to run the command it spawns grandchild processes.
So, we would need to find a way to not leave orphan processes.
On Unix/Linux we would use process groups on Windows there are job objects.
This commit avoids the problem all together by not using the shell anymore.